requirements.txt
numpy
Pillow
tensorflow
tensorflow_addons
streamlit